about and contact added
