+++ controllers/main.js
@@ -1,0 +1,11 @@
+exports.about=(req,res)=>{
+    res.render('./show/about');
+};
+exports.contact=(req,res)=>{
+    res.render('./show/contact');
+};
+
+
+
+
+
