+++ routes/main.js
@@ -1,11 +1,12 @@
 const express = require('express');
-const fs = require('fs');
+const mainController= require('../controllers/main');
 const router = express.Router();
-const app = express();
+
 //app.set('view engine', 'ejs');
 
 //GET 
-
+router.get('/about',mainController.about);
+router.get('/contact',mainController.contact);
