--- views/connection.ejs
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>iMeets</title>
-    <base href="/">
-    <!-- favicon -->
-    <link rel="icon" type="image/x-icon" href="assets/images/logo.png" />
-    <!-- Bootstrap cdn links -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN"
-      crossorigin="anonymous"
-    />
-    <script
-      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
-      integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
-      crossorigin="anonymous"
-    ></script>
-    <!-- header/footer styles -->
-    <link rel="stylesheet" href="styles/style.css" />
-    <!-- stylesheet -->
-    <link rel="stylesheet" href="styles/connectionStyle.css" />
-   
-  </head>
-
-  <body>
-    <!-- header section -->
-    <%- include("../partials/nav.ejs") %>
-    <!-- main content -->
-    <main class="container content-container">
-      <div class="jumbotron">
-        <h2>
-          <img id="eventPic" alt="" src="<%= connection.pic %>" /> &nbsp; <span id="eventTitle"><%= connection.title %></span>
-          <a href="../connections/connections.html" class="hrefColr"
-            ><button
-              class="btn btn-outline-success mx-lg-2 mx-0 my-lg-0 my-lg-2 me-auto ms-auto"
-              id="updateBut"
-            >
-              Update
-            </button></a
-          >
-          <a href="../connections/connections.html" class="hrefColr"
-            ><button
-              class="btn btn-outline-success mx-lg-2 mx-0 my-lg-0 my-lg-2 me-auto ms-auto"
-              id="deleteBut"
-            >
-              Delete
-            </button></a
-          >
-        </h2>
-        <br />
-        <br />
-        <table id="table1">
-          <tr>
-            <td><b>Presented By:</b></td>
-            <td id="hostedBy"><%= connection.host %></td>
-          </tr>
-          <tr>
-            <td><b>Venue:</b></td>
-            <td id="location"><%= connection.location %></td>
-          </tr>
-          <tr>
-            <td>
-              <b>Time:</b>
-            </td>
-            <td id="time"><%= connection.time %></td>
-          </tr>
-          <tr>
-            <td>
-              <b><u id="descTitle"><%= connection.descTitle %></u></b>
-            </td>
-          </tr>
-          <tr>
-            <td id="description"><%= connection.description %></td>
-          </tr>
-        </table>
-      </div>
-    </main>
-    <!-- footer section -->
-    <%- include("./partials/footer.ejs") %>
-  </body>
-</html>
