--- views/connections.ejs
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <title>iMeets</title>
-    <!-- favicon -->
-    <link rel="icon" type="image/x-icon" href="assets/images/logo.png" />
-    <!-- Bootstrap cdn links -->
-    <link
-    href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
-    rel="stylesheet"
-    integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN"
-    crossorigin="anonymous"
-  />
-  <script
-    src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
-    crossorigin="anonymous"
-  ></script>
-    <!-- stylesheet -->
-    <link rel="stylesheet" href="styles/style.css" />
-    <link rel="stylesheet" href="styles/connectionsStyle.css" />
-    <!-- scripts -->
-    <!-- <script src="./script.js" defer></script> -->
-  </head>
-
-  <body>
-    <!-- header section -->
-   <%- include("./partials/nav.ejs") %>
-    <!-- main content -->
-    <main class="container content-container">
-      <div class="jumbotron" id="listOfCategoriesContainer">
-       <% if(Object.keys(listOfConnections).length > 0) {%>
-       <h1>Categories in I-Meet</h1>
-       <% for(let category in listOfConnections) {%>
-        <h2> <%= category %></h2>
-        <ul class="lead">
-          <% for(let connection of listOfConnections[category]) {%>
-            <li>
-              <a href="<%= '/connection/'+connection.id %>"><%= connection.title %></a>
-            </li>
-          <% } %>
-        </ul>
-       <% } %>
-        <%  } else {%>
-          <p>No categories in I-Meet. Start a new <a href="/newConnection">connection.</a></p>
-          <% } %>
-      </div>
-    </main>
-    <!-- footer section -->
-    <%- include("./partials/footer.ejs") %>
-  </body>
-</html>
