--- views/newConnection.ejs
@@ -1,123 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>iMeets</title>
-    <!-- favicon -->
-    <link rel="icon" type="image/x-icon" href="assets/images/logo.png">
-    <!-- Bootstrap cdn links -->
-    <link
-      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
-      rel="stylesheet"
-      integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN"
-      crossorigin="anonymous"
-    />
-    <script
-      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
-      integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
-      crossorigin="anonymous"
-    ></script>
-    <!-- css files -->
-    <link rel="stylesheet" href="styles/style.css" />
-    <link rel="stylesheet" href="styles/newConnectionStyles.css" />
-    <!-- js files -->
-  </head>
-  <body>
-   <!-- header section -->
-    <%- include("./partials/nav.ejs") %>
-    <!-- main section -->
-    <div class="container-fluid">
-    <main class="new-connection-container">
-      <h1>Create a new connection</h1>
-      <form class="new-connection-form" action="./connections/connections.html">
-        <div class="mb-3 form-group">
-          <label for="topic" class="form-label">Topic</label>
-          <input
-            type="text"
-            class="form-control"
-            id="topic"
-            placeholder="Enter the topic of your connection"
-          />
-        </div>
-        <div class="mb-3 form-group">
-          <label for="title" class="form-label">Title</label>
-          <input
-            type="text"
-            class="form-control"
-            id="title"
-            placeholder="Enter the title of your connection"
-          />
-        </div>
-        <div class="mb-3 form-group">
-            <label for="host" class="form-label">Host</label>
-            <input
-              type="text"
-              class="form-control"
-              id="host"
-              placeholder="Enter the host of your connection"
-            />
-          </div>
-          <div class="mb-3 form-group">
-            <label for="details" class="form-label">Details</label>
-            <textarea              
-              class="form-control"
-              id="details"
-              placeholder="Enter the details of your connection"
-            ></textarea>
-          </div>
-          <div class="mb-3 form-group">
-            <label for="location" class="form-label">Where</label>
-            <input
-              type="text"
-              class="form-control"
-              id="location"
-              placeholder="Enter the location of your connection"
-            />
-          </div>
-          <div class="mb-3 form-group">
-            <label for="date" class="form-label">When</label>
-            <input
-              type="date"
-              class="form-control"
-              id="date"
-            />
-          </div>
-          <div class="mb-3 form-group">
-            <label for="start-time" class="form-label">Start</label>
-            <input
-              type="time"
-              class="form-control"
-              id="start-time"
-            />
-          </div>
-          <div class="mb-3 form-group">
-            <label for="end-time" class="form-label">End</label>
-            <input
-              type="time"
-              class="form-control"
-              id="end-time"
-            />
-          </div>
-          <div class="mb-3 form-group">
-            <label for="file" class="form-label">File</label>
-            <input
-              type="file"
-              class="form-control"
-              id="file"
-              accept="image/*"
-            />
-          </div>
-
-          <div class="d-grid gap-2 submit-btn-container">
-            <button type="submit" class="btn btn-primary">Create connection</button>
-          </div>
-       
-      </form>
-    </main>
-  </div>
-    <!-- footer section -->
-    <%- include("./partials/footer.ejs") %>
-  </body>
-</html>
