+++ views/partials/footer.ejs
@@ -5,7 +5,7 @@
    &nbsp;
  </p>
  <ul class="navbar-nav me-auto me-sm-2">
-   <li><a href="#" id="text4">ABOUT &nbsp;</a></li>
-   <li><a href="#" id="text5">CONTACT &nbsp;</a></li>
+   <li><a href="/main/about" id="text4">ABOUT &nbsp;</a></li>
+   <li><a href="/main/contact" id="text5">CONTACT &nbsp;</a></li>
  </ul>
 </footer>
