+++ views/show/about.ejs
@@ -1,0 +1,63 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>iMeets</title>
+    <base href="/">
+    <!-- favicon -->
+    <link rel="icon" type="image/x-icon" href="../images/logo.png" />
+    <!-- Bootstrap cdn links -->
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN"
+      crossorigin="anonymous"
+    />
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
+      crossorigin="anonymous"
+    ></script>
+    <!-- header/footer styles -->
+    <link rel="stylesheet" href="../styles/style.css" />
+    <!-- stylesheet -->
+    
+   
+  </head>
+
+  <body>
+    <!-- header section -->
+    <%- include("../partials/nav.ejs") %>
+    <main class="container content-container">
+        <div class="jumbotron">
+          <h2>About IMeets</h2>
+          <p class="lead">
+            Welcome to the world of IMeets, your passport to global meetups and unforgettable experiences.
+          </p>
+          <h2>Who We Are</h2>
+          <p class="lead">
+            At IMeets, we are passionate about connecting people and creating opportunities for meaningful interactions. We believe that great connections can lead to incredible experiences, and we've designed our app to make those connections easier and more accessible than ever before.
+          </p>
+          <h2>What We Do</h2>
+          <p class="lead">
+            IMeets is your go-to platform for discovering a wide range of events happening around the world. Whether you're looking for career-focused networking events, enriching arts and cultural gatherings, or simply want to explore diverse topics, IMeets has you covered. We provide you with valuable information about event organizers, venues, and timings, ensuring you never miss out on an opportunity that resonates with you.
+          </p>
+          <h2>Our Mission</h2>
+          <p class="lead">
+            Our mission is simple but powerful: to bring people together from all corners of the globe, fostering connections that lead to personal growth, shared experiences, and lasting relationships. We are committed to helping you explore your interests, expand your horizons, and make the most of every moment.
+          </p>
+          <h2>Join Us Today</h2>
+          <p class="lead">
+            Don't miss out on the chance to be part of a global community that values connections, experiences, and personal growth. Join the excitement on IMeets and embark on a journey of discovery, connection, and enrichment.
+          </p>
+
+          <br>
+          <br>
+        </div>
+      </main>
+
+
+    <%- include("../partials/footer.ejs") %>
+  </body>
+</html>
