+++ views/show/contact.ejs
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>iMeets</title>
+    <base href="/">
+    <!-- favicon -->
+    <link rel="icon" type="image/x-icon" href="../images/logo.png" />
+    <!-- Bootstrap cdn links -->
+    <link
+      href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css"
+      rel="stylesheet"
+      integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN"
+      crossorigin="anonymous"
+    />
+    <script
+      src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
+      integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
+      crossorigin="anonymous"
+    ></script>
+    <!-- header/footer styles -->
+    <link rel="stylesheet" href="../styles/style.css" />
+    <!-- stylesheet -->
+    
+   
+  </head>
+
+  <body>
+    <!-- header section -->
+    <%- include("../partials/nav.ejs") %>
+    <main class="container content-container">
+        <div class="jumbotron">
+          <h1>Contact</h1>
+         
+
+        </div>
+      </main>
+
+
+    <%- include("../partials/footer.ejs") %>
+  </body>
+</html>
